pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "bluesky-profile-sentiment-analysis"
version = "0.1.0"
description = "Perform sentiment analysis on the most recent posts from any BlueSky user"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "atproto>=0.0.62",
    "llama-index-workflows>=2.8.0",
    "nlpcloud>=1.1.47",
]

[tool.hatch.build.targets.wheel]
only-include = ["src/bluesky_profile_sentiment_analysis"]

[tool.hatch.build.targets.wheel.sources]
"src" = ""

[tool.llamadeploy.workflows]
sentiment-analysis = "bluesky_profile_sentiment_analysis.main:workflow"

[tool.llamadeploy]
name = "bluesky-profile-sentiment-analysis"
env_files = [".env"]
llama_cloud = true